1/3 - Major overhaul. Initial support for plan9
