+++ logos.go
@@ -1,0 +1,232 @@
+// This file contains all of the logos used within [truefetch](https://github.com/peajack/truefetch)
+package main
+
+func getLogo(id string) (Logo, bool) {
+	logos := map[string]Logo{
+		"arch": {
+			`      /\      `,
+			`     /  \     `,
+			`    /    \    `,
+			`   /      \   `,
+			`  /   ,,   \  `,
+			` /   |  |  -\ `,
+			`/_-''    ''-_\`,
+			`              `,
+			BCYAN,
+			"pacman",
+		},
+		"archbang": {
+			`          ____`,
+			`      /\ /   /`,
+			`     /  /   / `,
+			`    /   / /   `,
+			`   /   /_/\   `,
+			`  /   __   \  `,
+			` /   /_/\   \ `,
+			`/_-''    ''-_\`,
+			BCYAN,
+			"pacman",
+		},
+		"arcolinux": {
+			`              `,
+			`      /\      `,
+			`     /  \     `,
+			`    / /\ \    `,
+			`   / /  \ \   `,
+			`  / /    \ \  `,
+			` / / _____\ \ `,
+			"/_/  `----.\\_\\",
+			BBLUE,
+			"pacman",
+		},
+		"opensuse-leap": {
+			`  _______  `,
+			`__|   __ \ `,
+			`     / .\ \`,
+			`     \__/ |`,
+			`   _______|`,
+			`   \_______`,
+			`__________/`,
+			`           `,
+			BGREEN,
+			"rpm",
+		},
+		"debian": {
+			`  _____  `,
+			` /  __ \ `,
+			`|  /    |`,
+			`|  \\___-`,
+			`-_       `,
+			`  --_    `,
+			`         `,
+			`         `,
+			BRED,
+			"dpkg",
+		},
+		"fedora": {
+			`      _____   `,
+			`     /   __)\ `,
+			`     |  /  \ \`,
+			`  ___|  |__/ /`,
+			` / (_    _)_/ `,
+			`/ /  |  |     `,
+			`\ \__/  |     `,
+			` \ (_____/    `,
+			BBLUE,
+			"rpm",
+		},
+		"gentoo": {
+			`   _-----_   `,
+			`  (       \  `,
+			`  \    0   \ `,
+			`   \        )`,
+			`   /      _/ `,
+			`  (     _-   `,
+			`  \____-     `,
+			`             `,
+			BMAGENTA,
+			"portage",
+		},
+		"ubuntu": {
+			`           `,
+			`         _ `,
+			`     ---(_)`,
+			` _/  ---  \`,
+			`(_) |   |  `,
+			`  \  --- _/`,
+			`     ---(_)`,
+			`           `,
+			BRED,
+			"dpkg",
+		},
+		"linuxmint": {
+			` _____________ `,
+			`|_            \`,
+			` |  | _____  | `,
+			` |  | | | |  | `,
+			` |  | | | |  | `,
+			` |  \_____/  | `,
+			` \___________/ `,
+			`               `,
+			BGREEN,
+			"dpkg",
+		},
+		"manjaro": {
+			` ________  __ `,
+			`|       | |  |`,
+			`|   ____| |  |`,
+			`|  |  __  |  |`,
+			`|  | |  | |  |`,
+			`|  | |  | |  |`,
+			`|  | |  | |  |`,
+			`|__| |__| |__|`,
+			BGREEN,
+			"pacman",
+		},
+		"artix": {
+			`      /\      `,
+			`     /  \     `,
+			`    /''.,\    `,
+			`   /     ',   `,
+			`  /      ',\  `,
+			` /   ,.''.  \ `,
+			`/.,''     ''.\`,
+			`              `,
+			BCYAN,
+			"pacman",
+		},
+		"alpine": {
+			`    /\ /\    `,
+			`   /  \  \   `,
+			`  /    \  \  `,
+			` /      \  \ `,
+			`/        \  \`,
+			`          \  `,
+			`             `,
+			`             `,
+			BBLUE,
+			"apk",
+		},
+		"void": {
+			`      _____    `,
+			`   _  \____ -  `,
+			`  / / ____ \ \ `,
+			` / / /    \ \ \`,
+			` | |  VOID  | |`,
+			` \ \ \____/ / /`,
+			`  \ \____  /_/ `,
+			`   -_____\     `,
+			BGREEN,
+			"xbps",
+		},
+		"freebsd": {
+			`             `,
+			`/\,-'''''-,/\`,
+			`\_)       (_/`,
+			`|           |`,
+			`|           |`,
+			` ;         ; `,
+			`  '-_____-'  `,
+			`             `,
+			BRED,
+			"freebsd",
+		},
+		"openbsd": {
+			`      _____    `,
+			`    \-     -/  `,
+			` \_/         \ `,
+			` |        O O |`,
+			` |_  <   )  3 )`,
+			` / \         / `,
+			`    /-_____-\  `,
+			`               `,
+			BYELLOW,
+			"openbsd",
+		},
+		"netbsd":    {},
+		"dragonfly": {},
+		"ios":       {},
+		"macos":     {},
+		"plan9": {
+			`       `,
+			`  (\(\ `,
+			`¸". .. `,
+			`(  . .)`,
+			`|   ° ¡`,
+			`¿     ;`,
+			`c?".UJ"`,
+			`       `,
+			"",
+			"plan9",
+		},
+		"_UNKNOWN_": {
+			`     ___     `,
+			` ___/   \___ `,
+			`/   '---'   \`,
+			`--_______--' `,
+			`     / \     `,
+			`    /   \    `,
+			`   /     \   `,
+			`             `,
+			BWHITE,
+			"unknown",
+		},
+		"linux": {
+			`    .--.   `,
+			`   |o_o |  `,
+			`   |:_/ |  `,
+			`  /   \ \  `,
+			` (|     | )`,
+			`/'\_   _/'\`,
+			`\___)=(___/`,
+			`           `,
+			BWHITE,
+			"unknown",
+		},
+	}
+	// Compat: Support for different editions
+	logos["opensuse-tumbleweed"] = logos["opensuse-leap"]
+	// --------------------------------------
+	logo, present := logos[id]
+	return logo, present
+}
